Rename StartCall → Send.
Still playing with names a bit here; I think this is actually enough better to
be worth changing a second time. "Send" reflects the fact that it is not making
a call, but sending a bunch of requests -- some or all of which may in fact be
notifications, which never get replies.