Add NewCaller function.
This function supports generating RPC client code with a more natural interface
than is provided by the raw jrpc2.Client. Given representatives of the request
and response type, NewCaller reflectively constructs a function from the request
to the response via a client, i.e.,

   func(*jrpc2.Client, *Req) (Rsp, error)

This function encapsulates the call flow for a single invocation, using the
client to marshal and send the request message, waiting for the response, and
decoding the result (or error) into the return values.