Unexport some of the call machinery.
Unexport the Req and Send methods, and remove Note entirely.  The Batc, Call,
and Notify methods now call send directly and return the pending values for
their calls (or the error in the case of Notify).

This reduces the surface area of the API, at the cost of some flexibility.  I
think the value of having send exported was marginal.

Update package documentation for the API changes.