Remove the server.Proxy type.
I implemented this as a way of bridging to other framing disciplines, before
deciding to bake that into the Channel type and make it part of the core
server and client logic.

Now that it's done, this kluge is no longer necessary. Drop that part of the
package, its support code, and its test.