Make the server linger at shutdown for notifications.
When the connection is closed, do not exit immediately, but process any
remaining notifications that remain in the queue. Otherwise, a notification
from a client that disconnects immediately after sending a notification may be
lost. In general, it's fine for the server to keep processing after the channel
is closed; it just can't send back any replies.

During shutdown, after the channel is closed, walk through the queue and drop
any requests that have IDs (and thus want a response) but have not already been
processed. Leave notifications (having no ID).