Correctly report malformed parameters.
When a request message fails to decode because the paramters were not of the
proper shape (list or object), report this as an invalid request rather than
dropping the connection. We do this by passing back the saved ID in the error
data, which is arguably a little hacky but harmless.