Add server notification support to the *jrpc2.Server type.
If a server is constructed with the new AllowNotify option set true, its Notify
method will transmit a server-side notification to the client, and the context
passed to methods will carry a function to post notifications to the client
that can be retrieved by the ServerNotify helper.

If AllowNotify is false, Notify will return an error if called, and the method
context will not carry this value.

Server-side notifications are a non-standard extension to JSON-RPC.