Add server push to the example code.
Enable server push in the server, and add a hook to the client to receive
pushes. Make the Status method issue a server push back to the client and fail
if it doesn't work. This is just a way to verify that the plumbing works, and
to show off how to wire it up.