Handle reads from closed sockets and pipes better.
After closing a pipe or socket, a read will return an internal error whose text
contains "use of closed network connection". Unfortunately, that appears to be
the only way to detect that error, and https://github.com/golang/go/issues/4373
suggests it's not going to improve on its own.

Do what ~everyone else seems to have done in this case, check for the string.
That gives us nicer diagnostics when a client or server shuts down in the
normal course of its connections being terminated.