jctx: Add Authorizer and AuthToken.
The Encode function now checks for the presence of an Authorizer on the
outbound context. If one is found, it is used to generate a token that is
attached to the encoded wrapper.

The Decode function checks for a token on the wrapper and, if one is found,
attaches it to the inbound context. The AuthToken helper can be used to check
for the presence of a token and fetch its contents.