jrcp2: Reduce downward dependency on package jctx.
I mistakenly allowed a dependency from jrpc2 down into jctx. Rather than using
the jctx.AuthToken method directly, change the signature of the CheckAuth hook
to receive the context and the caller can choose how to get the token out of
the context.

For now the test still depends on the subpackage, but I'll move that in a
subsequent commit.