Tolerate params set to "null" as equivalent to unset.
The spec requires that parameters: "[i]f present, ... MUST be provided
... either by-position through an Array or by-name through an Object."
Strictly that means if you write "params": null, it's not valid, but the server
should tolerate that case since in effect it means the parameters are absent.

Silently treat "params": null as equivalent to unset parameters, and add a test
that confirms it makes a difference.