Split handler support into a separate package.
The NewHandler function, MapAssigner, and ServiceMapper types do not need to be
part of the main package. Move them to a separate package and rework the places
in the tests that required them to be combined.

- Move the Assigner and Handler definitions to base.go.
- Move most of the tests to package jrpc2_test.
- Update usage in subpackages.

A few of the tests need to examine unexported details of the implementation to
verify that the results are correct; move those tests to internal_test.go which
is still in the main jrpc2 package.