Ensure single-element batches retain their shape.
Although it does not specifically address this case, a close reading of the
JSON-RPC spec suggests that a batch consisting of a single valid request, not a
notification, should return an array. Previously the server did not do this:
Any single-element reply was returned as given.

With this change the server keeps track of whether each request came from a
batch, and when encoding the reply uses that to decide whether or not to return
a single element as an array.