Check and fail requests for extra fields.
By default the Go JSON decoder ignores unknown fields in a JSON object when
unmarshaling into a struct. Modify the custom decoder to check for this and
report it. The json.Decoder has a flag to disallow unknown fields, but we want
to keep track of request IDs for error reporting, so here I've made the loop
explicit.