Stop implementing json.Unmarshaler on *jrequests.
Without changing the functionality, avoid implementing the stock interface
because calling json.Unmarshal bypasses the custom handler if it detects that
the input message is structurally invalid. For this library, we want to handle
that case explicitly and control the error.

This commit just changes the name; the behaviour is the same.