Generalize the request checking interface.
Rename CheckAuth to CheckRequest, and remove code.NotAuthorized as a built-in
reserved value. The application may still define it, and this is exhibited in
the test. Although I originally implemented this for authorization, there is no
good reason to privilege that specific case, and it winds up leaving the API a
little smaller to remove that detail.