Preserve code values in their derived errors.
Prior to this change, the Err method of a code.Code value would return either
nil or a nonce error constructed by fmt.Errorf. But this is lossy, in that the
original code can no longer be recovered from that error value.

Instead, introduce an unexported wrapper type that preserves the code
value. The new type produces the same error string, but also implements the
Coder interface so the value can be recovered. The indirection preserves the
property that a Code is not itself an error (which we want so that we do not
accidentally use code.NoError where we want nil), but keeps the compact
representation.