handler: Document an unused method in the assignment test.
The staticcheck tool correctly flags this (U1000) as unused, since it is
unexported. But in this case the reason the method is there is to verify that
unexported methods are not picked up by the assignment builder, so that is as
intended. None of the methods on this type are "used".