channel: Add OptionalHeader framing
The existing `Header` framing does a few things, amongst others:
 1. attaches given Content-Type to outgoing requests
 2. checks that incoming requests have given Content-Type

In LSP, the Content-Type header is actually optional
and in reality many LSP clients choose to omit it, which makes
both LSP and Header Framing not fit for this reality.

Arguably we could change these clients to start sending the header,
but they are technically complying with the spec.

OptionalHeader fits this reality while offering the same functionality.