Allow header framed messages to omit Content-Type.
Rework the Header framing, moving the existing implementation to StrictHeader.
The latter still reports an error for content-type mismatches, but does so in a
way that permits the caller to detect it.

The Header framing is now a light wrapper for StrictHeader that filters out
these mismatches. The LSP framing now permits the sender to omit the
content-type field without reporting an error.

Addresses #4.