Remove the Stop method.
As implemented, this method was racy: A call to Stop invokes wait, which races
with the call to wait in Run on updating the service.

The motivation for Stop was to allow the service to stop the server, but
because the Service has to be constructed before the one-shot wrapper, it would
be ugly to plumb it in. If that turns out to be important, we can address it in
a different way.

This also allows the Run implementation to be simplified.