Consolidate protocol messages into a single type.
Updates #18. As a first step toward allowing a server to "push" calls to a
client, unify the jrequest and jresponse types into a single jmessage type that
represents protocol objects of either type. This has the pleasant side-effect
of simplifying the code for marshaling and unmarshaling as well.

No functional changes are intended, and all the tests still pass.