Slightly speed up task dispatch.
Prior to this change, a goroutine was started for each task in a batch, even if
there is only one. Since a goroutine is also started for the batch itself, this
is redundant and wasteful.

Handle this case by only starting goroutines when there are n > 1 tasks in the
batch, and then only for the first n - 1 of those tasks.

  BENCHMARK                           BEFORE   AFTER   SPEEDUP (%)
  BenchmarkRoundTrip/C01-CTX-B-12     21368    20092   6.0
  BenchmarkRoundTrip/C01-CTX+B-12     21411    20094   6.2
  BenchmarkRoundTrip/C04-CTX-B-12     21588    20055   7.1
  BenchmarkRoundTrip/C04-CTX+B-12     21375    20135   5.8
  BenchmarkRoundTrip/C12-CTX-B-12     21391    20020   6.4
  BenchmarkRoundTrip/C12-CTX+B-12     21475    20023   6.8
  BenchmarkRoundTrip/C01+CTX-B-12     24032    22516   6.3
  BenchmarkRoundTrip/C01+CTX+B-12     23871    22212   6.9
  BenchmarkRoundTrip/C04+CTX-B-12     23690    22097   6.7
  BenchmarkRoundTrip/C04+CTX+B-12     23633    22123   6.4
  BenchmarkRoundTrip/C12+CTX-B-12     23765    22248   6.4
  BenchmarkRoundTrip/C12+CTX+B-12     23589    22111   6.3