Add a barrier constraining execution order for notifications.
When issuing a batch of requests, keep track of the number that are valid
notifications, and do not issue to any handler for the batch until all
notifications received in previous batches have completed.

This ensures a sensible partial order on notifications. Multiple notifications
in a single batch can still be executed in parallel, but notifications across
batches are no longer concurrent.

Update tests and documentation.