Fix a spurious race in the metric tests.
The server fully transmits a response to the client before updating metrics. In
the unit test, this means the client can observe the request to complete before
the metrics are actually updated. This causes occasional spurious test failures
in the check for metrics being updated: Spurious, because the metric DOES get
updated, just not immediately.

Avert this issue by fully shutting down the server before doing any of the
metric checks. No functional changes.