Add support for server-side callbacks (#25)
Extend the server-side push mechanism to support calls back to the client
as well as notifications. This is a limited extension, without batch or context
support.

The server now exports a Callback method that makes a single, synchronous
call back to the client channel, blocking until a response is received. As with
notifications, this feature is gated by the AllowPush server option.

This commit includes a breaking API changes, specifically:

- The server method Push has been renamed Notify.
- The helper function ServerPush has been renamed PushNotify.
- The error value ErrNotifyUnsupported has been renamed ErrPushUnsupported.

Documentation and tests updated.

Co-authored-by: Radek Simko <radek.simko@gmail.com>