Allow non-string values for label metrics.
Prior to this change, labels can only contain strings. Now they can take any
JSON-marshalable value. This is technically a breaking change, but most
existing code should continue to compile -- correct usage of the old interface
is subsumed by the new. A caller that uses the SetLabel method as a callback
will need to be updated.