Fix request validation checks for push calls. (#37)
Prior to this change, the server checked the ID of an inbound request before
checking whether it is a callback response. This could result in a spurious
error if the callback happened to use the same ID as an in-flight request.

Update the tests to ensure they tickle this case, and fix the order of checks
so that we do not check for duplicate IDs until AFTER filtering matches of a
push response.

Fixes #36.