Handle panics in client-side callback handlers.
Prior to this change, a client callback handler that panics could cause a
server running in the same process to stall waiting for a reply, if the panic
was absorbed at a higher level without cleaning up the client.

Although this combination of conditions is less common, it can be tricky to
debug. To avert this, recover a panic from the callback handler and convert it
into errors back to the server.