jhttp: expose constructor options for an HTTP channel (#45)
Previously, the implementation hardwired the default HTTP client.
Add a new options type permitting the caller to replace the client.

This is a breaking change to the constructor signature. The existing
use within the module has been updated, but callers will need to be
updated upon upgrade. Existing use can be preserved by replacing

    jtthp.NewChannel(url)

with

    jhttp.NewChannel(url, nil)