error: Export fields and remove methods.
Rather than exposing trivial getter methods to fetch the fields of the Error
type, simply export the fields. This is simpler and more idiomatic.

This is a breaking change for code that extracts the *Error value.
In particular, uses of the Error type methods will need to be updated to remove
the call brackets, e.g.,

   e.Code()    ⇒ e.Code
   e.Message() ⇒ e.Message

See issue #46.