client: avoid writing to a closed channel when handling callbacks
Fixes #47.

Prior to this change, it is possible for the client channel to be closed, or
the client to be shut down, while the dispatcher is accepting a callback.
This could result in the client attempting to write through a nil channel
with a response from the callback.

To address the problem, check (while holding the lock) before dispatch:
If the client channel is closed, there is no point in dispatching a callback,
since there is no client to receive it. Notifications will still be dispatched,
since the handlers for those can still proceed.