Remove the non-standard built-in cancellation protocol (#50)
Since the addition of the OnCancel hook for the client, and the export of CancelRequest 
on the server, it is no longer necessary to have non-standard built-in logic for propagating
client-side cancellation to the server.

Remove the client-side support for sending rpc.cancel notifications, and also remove the
server-side magic for responding to it. Update tests to exercise calling back up to the server
from a cancellation hook.

This is a BREAKING CHANGE.

Fixes #49.