Reject structurally invalid request IDs.
The JSON-RPC 2.0 spec stipulates that request IDs should be strings or integral
number values. Prior to this commit, the library would accept basically any
JSON value as an ID. Now we reject objects (other than null), Booleans, and
arrays. The spec also says that fractional numbers should not be used, but we
still accept them.

Explicitly setting "id" to null is treated as equivalent to omitting the "id".