Update the Network helper to return a tuple.
Instead of returning only the inferred network type, return also the address
that the caller passed in. This allows a call to jrpc2.Network to be passed
directly in argument position to net.Listen, without having to echo the address
explicitly.

This is a breaking change for uses of the Network function, but does not affect
any other part of the module API.