handler: Drop the service.Service stub methods.
These aren't really serving any useful purpose. The server package can already
stub an Assigner trivially for a request loop, and the only other place where a
Service is applied is in server.Run, which is redundant for callers who do not
care about the exit status. Specifically, given

  server.Run(ch, svc, opts)

Can be replaced by:

  jrpc2.NewServer(assigner, opts).Start(ch).Wait()

if the service doesn't do anything useful with the Finish method.