channel: Remove the chanutil subpackage.
The only use of chanutil.Framing was in the jcall tool, so inline the
implementation there and remove the package from the module API.

This would be easy to reimplement if someone really needs it, but in real
applications you can't just swap the framing around willy-nilly and expect
things to work.  It's useful for jcall only because it's intended to be a
generic client.