handler: Fix a bug in the handling of pointer typed arguments
While reworking the interface in v0.21.2, I accidentally introduced a bug:
When the argument is of pointer type, I missed dropping the pointer when
constructing a fresh argument value, resulting in a pointer-to-pointer.

This was not caught by the existing tests, since it's a valid allocation, but
it would fail at runtime when attempting to unmarshal.

The fix is fortunately trivial, and I've added a regression test.

Also update module base version, add retractions; reduce CI versions.