Retract tag v0.23.0.
As reported in #53, I accidentally pushed two different commits for the v0.23.0
tag up to GitHub, and the first one got picked up by the module cache.
This version is safe to use, but the go tool will report a checksum error,
so I have marked it as retracted to avoid that friction.