jhttp: Reduce coordination between HTTP channel send/receive.
Each Send on an HTTP channel starts a goroutine to wait for the HTTP client to
finish executing and report its result. Previously, that goroutine could not
exit until the caller invoked Recv after finishing the HTTP request.

When the reply from the server is 204 (No content), it means the request is
complete and a reply is not pending -- this is what happens for notifications.
Previously, Recv filtered these out, looping until a non-empty response comes
back. But that means nothing happens until Recv is called.

Move this filtering logic from Recv to Send, so that the two only have to
rendezvous when there is a reply with "interesting" content (e.g., the response
or error from a call). This allows the Send goroutine to exit immediately
without waiting for Send in cases where that applies.

This change is not practically visible to the user, unless they were already
doing something unnatural.