handler: Document how to use arrays for single values.
Document that Check does not verify that the argument type encodes as an array
or object, and that JSON-RPC requires this. Explicitly note the recommended
solution (use a struct) and call out the simple array workaround for single
values with a different JSON encoding.

Add a test case to demonstrate use of an array-type parameter.