server: Fix a bug in duplicate request handling.
The server correctly detects duplicate requests and reports an error back to
the client. When dispatching results back to the client, however, it was using
the request ID to cancel its context: This means that reporting a duplicate
request ID could have the side effect of cancelling the in-flight (valid)
request with the same ID.

To avert this, only issue cancellations for request IDs in the batch that were
actually executed. Add a regression test for this bug.