server: Change the listener type for the Loop function.
Instead of taking a net.Listener, which is hard to implement outside the net
package because of the complexity of the net.Conn it returns, define a local
Accepter interface that returns a Channel directly. Since the Accepter is now
responsible for building the channel, the Framing option is no longer needed.

To make it easier to update existing use, add a NetAccepter function that
adapts a net.Listener and a framing to the new Accepter interface.

This is a breaking API change in the server package.