Support positional functions as handlers.
Add a new top-level function handler.Positional. Like Check, it accepts a
function to be wrapped as a jrpc2.Handler. Unlike Check, this function allows
positional arguments, which it implements by constructing a wrapper that takes
a synthetic struct type as its argument, and redirects the fields of the struct
to the positional arguments of the original function when the wrapper is called.