Rework the text logger interface.
Instead of using a standard log.Logger directly, define Logger as a function to
receive the formatted log string.

Add a helper (StdLogger) that adapts a standard logger to the new interface, to
simplify migration of existing use. Replace:

  Logger: log.New(...)

with:

  Logger: jrpc2.StdLogger(log.New(...))

to get the equivalent behaviour. The StdLogger function also accepts nil as an
argument to write to the default standard log.