server: Correctly terminate callbacks when shutting down.
If the Callback method is invoked with an arbitrary context, it may not
terminate when the server shuts down. The context passed to a handler will
correctly terminate pending callbacks, but the handler may create its own
context, and the server Callback can be invoked from outside a handler.

Handle this case by explicitly plumbing a cancellable context in to the
callback waiter, which will be used at shutdown even if the outer context does
not get triggered.