handler: support array arguments in positional wrappers
Extend the generated wrapper for a positional function to accept parameters
encoded as a JSON array of the correct length. When decoding the params, if
names are provided and the value is an array of the correct length, rewrite the
array into an equivalent object with the names mapped.

This works because the positional names preserve the original order, and ensure
that we do not introduce any unknown fields.