server: update context plumbing
Update the Accept method to accept a context. This is a breaking change to the
interface, but users of server.NetContext (which is most existing use) are not
affected.

Rework Loop as LoopContext, which now accepts a context that governs the loop.
Retain Loop as a wrapper to call LoopContext with a background context, so
existing use will work without change.