jhttp: Add optional Getter integration with a Bridge. (#69)
Add a field to BridgeOptions that enables GET requests to be handled by the
bridge using a Getter. Without this option, the behaviour of a Bridge does not
change. With it, "GET" requests are diverted to a Getter using the same server
and client options and the provided parse hook.