server: handle duplicate request IDs within a single batch
Previously, if two or more calls arrived in a single batch with the same
request ID, the server would process the first one and report the rest as
having duplicate IDs.

This is subtly wrong, in that the client has no way of knowing which of the
duplicates should get the successful call, and which should get the errors.
Although a client sending such requests is already wrong, this makes the error
harder to debug.

Instead, check for duplicates within the batch during assignment, and fail all
the colliding requests.

Add a regression test for duplicate requests in a single batch.