Release v0.42.0
This release removes the idiosyncratic "metrics" collector and replaces its use
with the standard "expvar" package (see #89). This is a breaking change to the
module API.

Packages that wish to attach custom metrics can now do so by calling the new
jrpc2.ServerMetrics helper, which returns a *expvar.Map. The caller may add any
metrics it wishes to this map, and they will be propagated into the ServerInfo
value by the server. The caller is also responsible for publishing the map to a
metrics collector (e.g., calling expvar.Publish or forwarding to Prometheus) if
desired.

Other changes in this release:
- Update module dependencies
- Remove usage of the "metrics" package from examples