client: improve handling of stopped clients
A client can be stopped by calling its Close method, or by a failure of its
connection to the server. The user of the client does not have a way to handle
this gracefully.

Add an IsStopped method to the *jrpc2.Client, allowing the caller to check
whether the client is still alive explicitly.

Add an optional OnStop callback that will be invoked when the client closes,
allowing the caller to handle cleanup tasks automatically.

Updates #102