jhttp: encode single-item batch responses as arrays in the Bridge
Although the JSON-RPC 2.0 spec does technically allow the responses from batch
requests to be returned separately, that is a peculiar corner and we shouldn't
rely on it.

Keep track of whether the original requests came in as a batch and use that
rather than response count to determine the encoding of the output from the
proxy.

Fixes #120