jrpc2: convert closures back into subtests
In a few places, move the synctest.Test plumbing inward so that we can use
subtests again rather than weird little closures.

In one of those places, we didn't really need synctest at all, so just remove
it entirely. I added it for the leak checking side-effect, but we are doing
that enough other places that this one doesn't matter.